react 훑어보기: 상품 정보 받아오기까지
