+++ .idea/.name
@@ -1,1 +1,1 @@
-gc-coffee
+package.json
