+++ src/main/java/org/programmers/gccoffee/configuration/MvcConfiguration.java
@@ -9,6 +9,6 @@ public class MvcConfiguration implements WebMvcConfigurer {
 
     @Override
     public void addCorsMappings(CorsRegistry registry) {
-        registry.addMapping("/api/&&").allowedOrigins("*");
+        registry.addMapping("/api/**").allowedOrigins("*");
     }
 }
